+++ examples/helloworld.py
@@ -3,14 +3,10 @@
 
 from lmnotify import LaMetricManager, Model, SimpleFrame
 
-# set your LaMetric API credentials here!
-CLIENT_ID = "<my_client_id>"
-CLIENT_SECRET = "<my_client_secret>"
-
 
 def main():
     # create an instance of the LaMetricManager
-    lmn = LaMetricManager(client_id=CLIENT_ID, client_secret=CLIENT_SECRET)
+    lmn = LaMetricManager()
 
     # get devices
     devices = lmn.get_devices()