+++ examples/notify.py
@@ -5,10 +5,6 @@ import argparse
 
 from lmnotify import LaMetricManager, Model, SimpleFrame
 
-# set your LaMetric API credentials here!
-CLIENT_ID = "<my_client_id>"
-CLIENT_SECRET = "<my_client_secret>"
-
 
 def main():
     # parse the command line arguments
@@ -24,7 +20,7 @@ def main():
     args = parser.parse_args()
 
     # create an instance of the LaMetricManager
-    lmn = LaMetricManager(client_id=CLIENT_ID, client_secret=CLIENT_SECRET)
+    lmn = LaMetricManager()
 
     # get devices
     devices = lmn.get_devices()