+++ examples/radio.py
@@ -5,14 +5,10 @@ import time
 
 from lmnotify import LaMetricManager
 
-# set your LaMetric API credentials here!
-CLIENT_ID = "<my_client_id>"
-CLIENT_SECRET = "<my_client_secret>"
-
 
 def main():
     # create an instance of the LaMetricManager
-    lmn = LaMetricManager(client_id=CLIENT_ID, client_secret=CLIENT_SECRET)
+    lmn = LaMetricManager()
 
     # get devices
     devices = lmn.get_devices()