+++ lmnotify/__init__.py
@@ -1,7 +1,8 @@
 __all__ = [
     "LaMetricManager", "SimpleFrame", "GoalFrame", "SpikeChart",
-    "Sound", "Model"
+    "Sound", "Model", "CloudSession", "LocalSession"
 ]
 
 from .lmnotify import LaMetricManager
 from .models import SimpleFrame, GoalFrame, SpikeChart, Sound, Model
+from .session import CloudSession, LocalSession