+++ CHANGELOG.rst
@@ -13,4 +13,15 @@ V0.0.3
 
 V0.0.4
 ^^^^^^
- * enabled authentication via constructor, environment variables and config file
+ * enabled authentication via constructor, environment variables 
+   and config file
+
+V0.0.6
+^^^^^^
+ * added local sessions and cloud sessions so that all functions that
+   can be accessed from the local network make use of the local session 
+ * added setting of screensaver via 'set_screensaver' 
+   (supported by firmware 2.10)
+ * added logging for easier debugging
+ * some additional fixes
+ * improved docstrings of main class